Updating the keygen code
